+++ src/components/menu/mobile.tsx
@@ -10,7 +10,7 @@ import {
   AiOutlineSetting,
 } from "react-icons/ai";
 import { MdOutlineLightMode, MdDarkMode } from "react-icons/md";
-import { HiLightBulb, HiOutlineTranslate } from "react-icons/hi";
+import { HiOutlineTranslate } from "react-icons/hi";
 import { RiFeedbackLine } from "react-icons/ri";
 import { v4 as uuidv4 } from "uuid";
 import { cn } from "@/lib";
@@ -113,8 +113,6 @@ const MobileMenu: React.FC = () => {
   };
 
   const onToggleTheme = () => setTheme(nowTheme === "light" ? "dark" : "light");
-
-  const onOpenPrompt = () => alert("Prompt Manage ToDo...");
 
   const onLocaleChange = (value: string) => {
     if (value === locale) return;
@@ -283,18 +281,6 @@ const MobileMenu: React.FC = () => {
               >
                 <AiFillGithub size={20} />
               </a>
-            </div>
-            <div className="flex-1 flex justify-center">
-              <div
-                onClick={onOpenPrompt}
-                className={cn(
-                  "w-8 h-8 flex justify-center items-center cursor-pointer transition-colors rounded-md",
-                  "hover:bg-gray-200/60",
-                  "dark:hover:bg-slate-700/70"
-                )}
-              >
-                <HiLightBulb size={20} />
-              </div>
             </div>
             <Dropdown
               selectable